cli: simplify error handling with just using `ArgumentError`
