+++ compiler/hash-pipeline/src/lib.rs
@@ -5,7 +5,7 @@
 //! parser or typechecker and just use a common trait interface that can be
 //! used. This file also has definitions for how to access sources whether
 //! module or interactive.
-pub mod error;
+pub mod args;
 pub mod fs;
 pub mod interface;
 pub mod settings;
