+++ compiler/hash-pipeline/src/settings.rs
@@ -3,9 +3,9 @@
 //! to the Compiler pipeline.
 use std::{fmt::Display, str::FromStr};
 
-use hash_target::{data_layout::TargetDataLayout, Target, TargetInfo};
+use hash_target::{data_layout::TargetDataLayout, TargetInfo};
 
-use crate::error::{ArgumentError, CompilerError};
+use crate::args::ArgumentError;
 
 /// Various settings that are present on the compiler pipeline when initially
 /// launching.
@@ -142,7 +142,7 @@ pub enum OptimisationLevel {
 }
 
 impl FromStr for OptimisationLevel {
-    type Err = CompilerError;
+    type Err = ArgumentError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
@@ -281,166 +281,4 @@ impl Display for CompilerStageKind {
             CompilerStageKind::Full => write!(f, "total"),
         }
     }
-}
-
-/// This function is used to parse the command line arguments that are
-/// passed to the compiler, it will return a [CompilerSettings] struct
-/// that contains all of the settings that the compiler should use. If
-/// there is an error, this will return an error that can be
-/// dealt with by the caller.
-pub fn parse_settings_from_args() -> Result<CompilerSettings, CompilerError> {
-    let mut settings = CompilerSettings::default();
-    let mut args = std::env::args().skip(1);
-
-    while let Some(arg) = args.next() {
-        // This is a configuration key that specifies the "key" and then
-        // the value in the form of `-C<key>=<value>`
-        if arg.starts_with("-C") {
-            // Check if the key is part of this argument or if there is a
-            // separate argument for the value, which we assume to be the
-            // next argument.
-            if arg.len() == 2 {
-                if let Some(arg) = args.next() {
-                    parse_arg_configuration(&mut settings, arg)?;
-                } else {
-                    return Err(CompilerError::Argument(ArgumentError::UnknownKey(arg)));
-                }
-            } else {
-                parse_arg_configuration(&mut settings, arg.trim_start_matches("-C").to_string())?;
-            }
-        } else if arg.starts_with("--") {
-            let key = arg.trim_start_matches("--").to_string();
-
-            match key.as_str() {
-                "debug" => {
-                    settings.debug = true;
-                }
-                "output-metrics" => {
-                    settings.output_metrics = true;
-                }
-                _ => {
-                    return Err(CompilerError::Argument(ArgumentError::UnknownKey(arg)));
-                }
-            }
-        } else {
-            // This is specifying what kind of a stage the compiler should run
-            // the job on whether it is `build`, `check`,
-            // `ast-gen`...
-            match arg.as_str() {
-                "build" => {
-                    settings.stage = CompilerStageKind::Full;
-                }
-                "check" => {
-                    settings.stage = CompilerStageKind::Typecheck;
-                }
-                "ast-gen" => {
-                    settings.stage = CompilerStageKind::Parse;
-                }
-                "ir-gen" => {
-                    settings.stage = CompilerStageKind::IrGen;
-                }
-                _ => {
-                    return Err(CompilerError::UnknownStage(arg));
-                }
-            };
-
-            // The next argument after this is the input file.
-            if let Some(filename) = args.next() {
-                settings.entry_point = Some(filename);
-            } else {
-                return Err(CompilerError::MissingEntryPoint);
-            }
-        }
-    }
-
-    Ok(settings)
-}
-
-/// This function will parse a single `-C<key>=<value>` argument and apply the
-/// specified configuration option to the [CompilerSettings]. Some keys may
-/// not have a value, and some keys may specify multiple values with a comma
-/// separated list.
-fn parse_arg_configuration(
-    settings: &mut CompilerSettings,
-    arg: String,
-) -> Result<(), CompilerError> {
-    // First try and see if we have been provided a key-value pair, if not
-    // then we will assume that the key is the argument and the value is
-    // `None`.
-    let (key, value) = if let Some(split_pos) = arg.find('=') {
-        let (key, value) = arg.split_at(split_pos);
-        (key.to_string(), Some(value[1..].to_string()))
-    } else {
-        (arg.clone(), None)
-    };
-
-    // When a value is expected from a key, but none is provided, this
-    // closure will be used to return an error.
-    let expected_value = || CompilerError::Argument(ArgumentError::MissingValue(key.clone()));
-
-    match key.as_str() {
-        "target" => {
-            let value = value.ok_or_else(expected_value)?;
-
-            let target = Target::from_string(value.clone())
-                .ok_or_else(|| CompilerError::InvalidTarget(value))?;
-
-            settings.codegen_settings.target_info.set_target(target)
-        }
-        "optimisation-level" => {
-            let value = value.ok_or_else(expected_value)?;
-            let opt_level = OptimisationLevel::from_str(value.as_str())?;
-            settings.optimisation_level = opt_level;
-        }
-        "dump" => {
-            let value = value.ok_or_else(expected_value)?;
-
-            match value.as_str() {
-                "ast" => {
-                    settings.ast_settings.dump = true;
-                }
-                "ir" => {
-                    settings.lowering_settings.dump = true;
-                }
-                _ => {
-                    return Err(CompilerError::Argument(ArgumentError::InvalidValue(key, value)));
-                }
-            }
-        }
-        "ir-dump-mode" => {
-            let value = value.ok_or_else(expected_value)?;
-
-            match value.as_str() {
-                "pretty" => {
-                    settings.lowering_settings.dump_mode = IrDumpMode::Pretty;
-                }
-                "graph" => {
-                    settings.lowering_settings.dump_mode = IrDumpMode::Graph;
-                }
-                _ => {
-                    return Err(CompilerError::Argument(ArgumentError::InvalidValue(key, value)));
-                }
-            }
-        }
-        "backend" => {
-            let value = value.ok_or_else(expected_value)?;
-
-            match value.as_str() {
-                "llvm" => {
-                    settings.codegen_settings.backend = CodeGenBackend::LLVM;
-                }
-                "vm" => {
-                    settings.codegen_settings.backend = CodeGenBackend::VM;
-                }
-                _ => {
-                    return Err(CompilerError::Argument(ArgumentError::InvalidValue(key, value)));
-                }
-            }
-        }
-        _ => {
-            return Err(CompilerError::Argument(ArgumentError::UnknownKey(key)));
-        }
-    };
-
-    Ok(())
 }
